+++ plugs/ui/js/jsutils/static/jsutils/desktopnotify.js
@@ -13,9 +13,23 @@ var Notify = {
     
     // Behavior Config. params
     autoclose           : null,
+    
+    // internal variables
+    _type : '',
 
     isSupported : function () {
-        return window.webkitNotifications;
+        var n;
+        if (window.webkitNotifications)
+        {
+            n = window.webkitNotifications;
+            this._type = 'webkit';
+        }
+        else if (window.Notification)
+        {
+            n = window.Notification;
+            this._type = 'w3c';
+        }
+        return n;
     },
 
     log: function(){
@@ -26,12 +40,26 @@ var Notify = {
     },
     request: function(){
         var notify = this.isSupported();
-        if (notify.checkPermission() == 1)
+        if (this.check() == 1)
             notify.requestPermission();                     
     },
     check: function(){
         var notify = this.isSupported();
-        return notify.checkPermission();
+        if (notify){
+            if (this._type == 'webkit')
+                return notify.checkPermission();
+            else{
+                var level;
+                
+                /* firefox is permission, it's not a function */
+                if (notify.permission) level = notify.permission;
+                else level = notify.permissionLevel();
+
+                if (level == 'granted') return 0;
+                else if (level == 'default') return 1;
+                else return 3;
+            }
+        }
     },
     /* ------------------------------------------
         Shows the notification (if possible)
@@ -39,7 +67,7 @@ var Notify = {
             icon:
             title:
             message:
-            ondisplay:
+            onshow:
             onclick:
             onerror:
             onclose:
@@ -51,13 +79,13 @@ var Notify = {
             icon: message.icon || this.icon,
             title: message.title || this.title,
             message: message.message || this.message,
-            ondisplay: message.ondisplay || function(){},
+            onshow: message.onshow || function(){},
             onclick: message.onclick || function(){},
             onerror: message.onerror || function(){},
             onclose: message.onclose || function(){}
         };
+        var check, n;
         
-        this.log('test');
         var notify = this.isSupported();
         //If webkitNotifications object is not available and we are in debug mode, an exception will be thrown...
         if (!notify) {
@@ -66,19 +94,28 @@ var Notify = {
         else {
             //Check for permission to show notifications. Request permission if notifications are not allowed
             this.request();
-            if (notify.checkPermission() == 2) {
-                this.log("Permission are denied!!" + notify.checkPermission());
+            check = this.check();
+
+            if (check == 2) {
+                this.log("Permission are denied!!");
             }
             //If permission is allowed, the notification is shown.
-
-            if (notify.checkPermission() == 0) {
-                var n = notify.createNotification(opts.icon, opts.title, opts.message);     
-                n.ondisplay = opts.ondisplay;
+            if (check == 0) {
+                if (this._type == 'webkit')
+                {
+                    n = notify.createNotification(opts.icon, opts.title, opts.message);     
+                    n.ondisplay = opts.onshow;
+                }
+                else {
+                    n = new notify(opts.title, {body:opts.message, icon:opts.icon});
+                    n.onshow = opts.onshow;
+                }
                 n.onclick = opts.onclick;
                 n.onerror = opts.onerror;
                 n.onclose = opts.onclose;
-
-                n.show();
+                
+                if (this._type == 'webkit')
+                    n.show();
 
                 if (this.autoclose) 
                     setTimeout(function () {